--- src/pages/HomePage/types.ts
@@ -1,7 +1,0 @@
-import { SlsDbItem } from '../../types';
-
-export type Skill = SlsDbItem & {
-  alt: string;
-  id: number;
-  url: string;
-};
