--- src/pages/TeamPage/types.ts
@@ -1,9 +1,0 @@
-import { SlsDbItem } from '../../types';
-
-export type TeamMember = SlsDbItem & {
-  name: string;
-  role: string;
-  profileImage?: string;
-  gif?: string;
-  alt?: string;
-};
