+++ src/types/types.ts
@@ -4,3 +4,11 @@ export type SlsDbItem = {
   created: string;
   value?: Record<string, any>;
 };
+
+export type TeamMember = SlsDbItem & {
+  name: string;
+  role: string;
+  profileImage?: string;
+  gif?: string;
+  alt?: string;
+};
